+++ lib/graphql.dart
@@ -1,7 +1,5 @@
 library graphql;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export './src/client.dart';
+export './src/query.dart';
+export './src/mutation.dart';