+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - June 15 2018.
 
-* TODO: Describe initial release.
+My collegue and I created a sipmple implementation of a GraphQL Client for Flutter. (Many thanks to Eus Dima, for his work on the innitial client.)
+
+- A client to connect to your GraphQL server.
+- A query widget to handle GraphQL queries.
+- A mutation widget to handle GraphQL mutations.
+- Simple support for query polling.